backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
flask-sqlalchemy==3.0.5
flask-bcrypt==1.0.1
flask-wtf==1.1.1
wtforms==3.0.1
bcrypt==4.0.1
tensorflow==2.20.0  # ✅ Working with Python 3.13
pillow==10.4.0
numpy>=2.2.6  # ✅ Compatible with TensorFlow 2.20.0
python-dotenv==1.0.0
flask-mail==0.9.1
openai==1.3.0
gunicorn==21.2.0
opencv-python-headless>=4.12.0  # ✅ Working image processing
matplotlib  # ✅ Working visualization
scipy  # ✅ Working scientific computing
werkzeug==2.3.7

# New dependencies for comprehensive features
googlemaps==4.10.0  # Google Maps integration
twilio==8.10.0  # SMS notifications
cloudinary==1.36.0  # Image storage and management
redis==5.0.1  # Caching and sessions
celery==5.3.4  # Background job processing
flask-socketio==5.3.6  # Real-time features
python-socketio==5.10.0  # Socket.io support
requests==2.31.0  # HTTP requests
pytesseract==0.3.10  # OCR for prescription images
cryptography==41.0.8  # Encryption for sensitive data
flask-limiter==3.5.0  # Rate limiting
sqlalchemy-utils==0.41.1  # Database utilities
